refactor(GameManager): 플레이어의 수를 받아오는 메서드 returnPlayerNumber() 구현
