+++ src/main/java/baseball/view/InputHandler.java
@@ -4,10 +4,9 @@ package baseball.view;
 import camp.nextstep.edu.missionutils.Console;
 
 public class InputHandler {
-    public int insertIntNum(){
+    public String insertBallNum(){
         System.out.print("숫자를 입력해주세요 : ");
-        String inputString = Console.readLine();
-        return Integer.parseInt(inputString);
+        return Console.readLine();
     }
 
 }
